Refactored classes into packages
