+++ src/com/jboby93/markovbot/Tools.java
@@ -8,6 +8,7 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.TimeZone;
 
+import logging.Logger;
 import sun.management.VMManagement;
 
 public class Tools {
